+++ NEWS.md
@@ -1,3 +1,10 @@
+# Changements v 0.1.2
+
+* Meilleure doc pour transform_gbif_parquet.R. 
+* Ajout de compression et level pour passer à la copie parquet. 
+* join_gbif_admin peut filtrer dynamiquement. Sélection des
+filtres pour certaines colonnes dont les kindoms et l'incertitude.
+
 # Changements v 0.1.1
 
 * Ajout de fonction pour télécharger les données de hotspots eBird